// frontend/src/api/generated/models/attributeOptionDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */
import type { AttributeOptionDetailedDto } from './attributeOptionDetailedDto';

export interface AttributeOptionDto {
  id: number;
  mlId?: string;
  valueName: string;
  attributeId: string;
  attributeName: string;
  detailed?: AttributeOptionDetailedDto;
}