// frontend/src/api/generated/models/variationDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */
import type { AttributeOptionDto } from './attributeOptionDto';
import type { PictureDto } from './pictureDto';

export interface VariationDto {
  id: string;
  price: number;
  availableQuantity: number;
  soldQuantity: number;
  attributeOptions: AttributeOptionDto[];
  pictures: PictureDto[];
  /** @nullable */
  variantName?: string | null;
}